unify unix file handling
i.e. anything that doesnt understand `echo -n` for speedup
