+++ lib/train/extras.rb
@@ -6,8 +6,8 @@ module Train::Extras
   autoload :CommandWrapper, 'train/extras/command_wrapper'
   autoload :FileCommon,     'train/extras/file_common'
   autoload :AixFile,        'train/extras/file_aix'
+  autoload :UnixFile,       'train/extras/file_unix'
   autoload :LinuxFile,      'train/extras/file_linux'
-  autoload :SolarisFile,    'train/extras/file_solaris'
   autoload :WindowsFile,    'train/extras/file_windows'
   autoload :OSCommon,       'train/extras/os_common'
   autoload :Stat,           'train/extras/stat'
